Added Email functionality.
Details on slack.

+++ api/KnowledgeBase.php
@@ -1,12 +1,12 @@
 <?php
 
-	function getFiles($id, $courseID, $fileName, $tutorserver) 
+	function getFiles($id, $courseID, $fileName, $tutorserver)
 	{
-		$query = "SELECT * FROM KnowledgeFiles WHERE 
+		$query = "SELECT * FROM KnowledgeFiles WHERE
 			ID=COALESCE(?, ID) AND
 			courseID=COALESCE(?, courseID) AND
 			fileName=COALESCE(?, fileName)";
-        
+
 		$files = array();
 		if($stmnt = $tutorserver->prepare($query)) {
             $stmnt->bind_param('iis', $id, $courseID, $fileName);
@@ -19,10 +19,10 @@
             return $files;
         }
     }
-	
+
 	function addFile($courseID, $userID, $fileName, $content, $tutorserver)
 	{
-		$query = "INSERT INTO KnowledgeFiles (courseID, userID, fileName, content) VALUES (?, ?)";
+		$query = "INSERT INTO KnowledgeFiles (courseID, userID, fileName, content) VALUES (?, ?, ?, ?)";
 		//for each ? bind parameter as seen below
         if($stmnt = $tutorserver->prepare($query)) {
             $stmnt->bind_param('iiss',
@@ -35,7 +35,7 @@
             $stmnt->close();
         }
 	}
-	
+
 	function setFile($fileID, $courseID, $userID, $fileName, $content, $approved, $tutorserver){
         $query = "UPDATE KnowledgeFiles SET
             courseID    =COALESCE(?, courseID),
@@ -51,13 +51,13 @@
             $stmnt->close();
         }
 	}
-	
-	function removeFile($fileID, $tutorserver)
+
+	function removeFile($userID, $fileID, $tutorserver)
 	{
-		$query = "DELETE FROM KnowledgeFiles WHERE ID=?";
-		
+		$query = "DELETE FROM KnowledgeFiles WHERE ID=? AND userID=?";
+
 		if($stmnt = $tutorserver->prepare($query)) {
-            $stmnt->bind_param('i', $fileID);
+            $stmnt->bind_param('ii', $fileID, $userID);
             $stmnt->execute() or trigger_error($stmt->error, E_USER_ERROR);
             $stmnt->close();
         }
